Ball and Player classes moved to separate files. Added logic to detect when ball is out of playfield.
